Inserted all the monkey patches from graphingwiki.__init__ with their dependencies to MoinMoin files. Started to structure a MoinMoin.metadata hierarchy. Code is not yet testable, but getting there.
