+++ MoinMoin/formatter/nullformatter.py
