+++ graphingwiki/__init__.py
@@ -23,8 +23,6 @@ import socket
 import xmlrpclib
 from cStringIO import StringIO
 
-SEPARATOR = '-gwikiseparator-'
-
 def RequestCLI(pagename='', parse=True):
     """
     The MoinScript class does command line argument parsing, which
@@ -236,200 +234,6 @@ class MoinRemoteWikiHttpAuth(MoinMoin.wikisync.MoinRemoteWiki):
 
 MoinMoin.wikisync.MoinRemoteWiki = MoinRemoteWikiHttpAuth
 
-# Helper functions for monkey patching and dealing with underlays.
-
-def ignore(*args, **keys):
-    pass
-
-def monkey_patch(original, on_success=ignore, always=ignore):
-    def _patched(self, *args, **keys):
-        try:
-            result = original(self, *args, **keys)
-        finally:
-            always(self)
-        # If we want to patch the result also
-        patched_result = on_success(self, result, (args, keys))
-        if patched_result:
-            return patched_result
-        return result
-    return _patched
-
-def underlay_to_pages(req, p):
-    underlaydir = req.cfg.data_underlay_dir
-
-    pagepath = p.getPagePath()
-
-    # If the page has not been created yet, create its directory and
-    # save the stuff there
-    if underlaydir in pagepath:
-        pagepath = pagepath.replace(underlaydir, pagepath)
-        if not os.path.exists(pagepath):
-            os.makedirs(pagepath)
-
-    return pagepath
-
-# Functions for properly opening, closing, saving and deleting
-# graphdata. NB: Do not return anything in functions used in
-# monkey_patch unless you want to affect the return value of the
-# patched function
-
-def graphdata_getter(self):
-#    from graphingwiki.backend.couchdbclient import GraphData
-#    from graphingwiki.backend.durusclient import GraphData
-    from graphingwiki.backend.shelvedb import GraphData
-    if "_graphdata" not in self.__dict__:
-        dbconfig = getattr(self.cfg, 'dbconfig', {})
-        if "dbname" not in dbconfig:
-            dbconfig["dbname"] = self.cfg.interwikiname
-
-        self.__dict__["_graphdata"] = GraphData(self, **dbconfig)
-    self.__dict__["_graphdata"].doing_rehash = _is_rehashing
-    return self.__dict__["_graphdata"]
-
-def graphdata_close(self):
-    graphdata = self.__dict__.pop("_graphdata", None)
-    if graphdata is not None:
-        graphdata.commit()
-        graphdata.close()
-
-def graphdata_commit(self, *args):
-    graphdata = self.__dict__.pop("_graphdata", None)
-    if graphdata is not None:
-        graphdata.commit()
-
-def _get_save_plugin(self):
-    # Save to graph file if plugin available.
-    try:
-        graphsaver = importPlugin(self.request.cfg, "action", "savegraphdata")
-    except PluginMissingError:
-        return
-
-    return graphsaver
-
-## TODO: Hook PageEditor.sendEditor to add data on template to the
-## text of the saved page?
-
-def graphdata_save(self, result, _):
-    graphsaver = _get_save_plugin(self)
-
-    if not graphsaver:
-        return
-
-    path = underlay_to_pages(self.request, self)
-    text = self.get_raw_body()
-
-    graphsaver(self.page_name, self.request, text, path, self)
-
-def graphdata_copy(self, result, (args, _)):
-    newpagename = args[0]
-    graphsaver = _get_save_plugin(self)
-
-    if not graphsaver:
-        return
-
-    text = self.get_raw_body()
-    path = underlay_to_pages(self.request, self)
-
-    graphsaver(newpagename, self.request, text, path, self)
-
-def graphdata_rename(self, (success, msg), _):
-    if not success:
-        return
-
-    graphsaver = _get_save_plugin(self)
-    path = underlay_to_pages(self.request, self)
-
-    # Rename is really filesystem-level rename, no old data is really
-    # left behind, so it should be cleared.  When saving with text
-    # 'deleted\n', no graph data is actually saved.
-    graphsaver(self.page_name, self.request, 'deleted\n', path, self)
-
-    # Rename might litter empty directories data/pagename and
-    # data/pagename/cache, let's remove them
-    oldpath = self.getPagePath(check_create=0)
-    for dirpath, dirs, files in os.walk(oldpath, topdown=False):
-        # If there are files left, some backups etc information is
-        # still there, so let's quit
-        if files:
-            break
-
-        os.rmdir(dirpath)
-
-def variable_insert(self, result, _):
-    """
-    Replace variables specified in wikiconfig eg.
-
-    gwikivariables = {'GWIKITEST': 'I am the bestest!'}
-    """
-
-    cfgvar = getattr(self.request.cfg, 'gwikivariables', dict())
-    for name in cfgvar:
-        result = result.replace('@%s@' % name, cfgvar[name])
-
-    form = self.request.values.to_dict(flat=False)
-
-    # Add the page's creator as a dynamic variable
-    backto = form.get('backto', [''])[0]
-    result = result.replace('@CREATORPAGE@', backto)
-
-    return result
-
-def acl_user_expand(self, result, _):
-    """
-    Expand @ME@ user variable, which can result into problems when
-    viewing template pages with acl:s
-    """
-    modifier, entries, rights = result
-
-    entries = [x.replace('@ME@', 'All') for x in entries]
-
-    return (modifier, entries, rights)
-
-
-def attachfile_filelist(self, result, (args, _)):
-    _ = self.getText
-    attachments = re.findall('do=view&amp;target=([^"]+)', result)
-    if len(attachments) < 2:
-        return result
-
-    form = u'<form method="GET" action="%s">\n' % \
-        actionname(self, self.page.page_name) + \
-        u'<input type=hidden name=action value="AttachFile">'
-
-    result = form + result
-
-    oldform = self.values.to_dict(flat=False)
-
-    att1 = oldform.get('att1', [''])[0]
-    att2 = oldform.get('att2', [''])[0]
-    sort = oldform.get('sort', ['normal'])[0]
-
-    for target in attachments:
-        buttontext = '\\1 | ' + \
-            '<input type="radio" value="%s" name="att1"/%s>' % \
-            (target, att1 == target and ' checked' or '')+ \
-            '<input type="radio" value="%s" name="att2"/%s>' % \
-            (target, att2 == target and ' checked' or '')+ \
-            _('diff')
-
-        viewtext = '(<a href.+&amp;do=view&amp;target=%s">%s</a>)' % \
-            (re.escape(target), _("view"))
-        
-        result, count = re.subn(viewtext, buttontext, result, 1)
-
-    result = result + \
-        '<input type="radio" value="normal" name="sort"/%s>%s\n' % \
-        (sort == 'normal' and ' checked' or '', _("Normal")) + \
-        '<input type="radio" value="sort" name="sort"/%s>%s\n' % \
-        (sort == 'sort' and ' checked' or '', _("Sort")) + \
-        '<input type="radio" value="uniq" name="sort"/%s>%s\n' % \
-        (sort == 'uniq' and ' checked' or '', _("Sort + uniq")) + \
-        '<input type="radio" value="cnt" name="sort"/%s>%s\n' % \
-        (sort == 'cnt' and ' checked' or '', _("Sort + uniq + count")) + \
-        '<br><input type=submit name=do value="diff"></form>'
-
-    return result
-
 # Main function for injecting graphingwiki extensions straight into
 # Moin's beating heart.
 
@@ -440,64 +244,6 @@ def install_hooks(rehashing=False):
     if _hooks_installed:
         return
     _is_rehashing = rehashing
-
-    # Monkey patch the request class to have the property "graphdata"
-    # which, if used, is then closed properly when the request
-    # finishes.
-    MoinMoin.web.contexts.Context.graphdata = property(graphdata_getter)
-
-    # Monkey patch the different saving methods to update the metas in
-    # the meta database.
-    # Note: PageEditor.renamePage seems to use .saveText for the new
-    # page (thus already updating the page's metas), so only the old page's
-    # metas need to be deleted explicitly.
-    PageEditor.saveText = monkey_patch(PageEditor.saveText, 
-                                       graphdata_save)
-    PageEditor.renamePage = monkey_patch(PageEditor.renamePage, 
-                                         graphdata_rename)
-    PageEditor.copyPage = monkey_patch(PageEditor.copyPage, 
-                                       graphdata_copy)
-
-    # Monkey patch AttachFile to include diffing
-    AttachFile._build_filelist = monkey_patch(AttachFile._build_filelist, 
-                                              attachfile_filelist)
-
-    # FIXME: Remove this patch when MoinMoin makes variable names
-    # configurable in some fashion.
-    PageEditor._expand_variables = monkey_patch(PageEditor._expand_variables,
-                                                variable_insert)
-
-    # Patch wsgiapp.run to close graphdb to avoid db corruption with
-    # some backends
-    orig_wsgiapp_run = MoinMoin.wsgiapp.run
-    def patched_run(context):
-        if not hasattr(context, '_patched'):
-            context.finish = lambda: graphdata_close(context)
-            context._patched = True
-        return orig_wsgiapp_run(context)
-    MoinMoin.wsgiapp.run = patched_run
-
-    # XMLRPC actions need to be patched separately, as wsgiapp.run
-    # makes a new context for them
-    orig_xmlrpc2 = MoinMoin.xmlrpc.xmlrpc2
-    def patched_xmlrpc2(context):
-        try:
-            return orig_xmlrpc2(context)
-        finally:
-            graphdata_close(context)
-    MoinMoin.xmlrpc.xmlrpc2 = patched_xmlrpc2
-
-    orig_xmlrpc = MoinMoin.xmlrpc.xmlrpc
-    def patched_xmlrpc(context):
-        try:
-            return orig_xmlrpc(context)
-        finally:
-            graphdata_close(context)
-    MoinMoin.xmlrpc.xmlrpc = patched_xmlrpc
-
-    # Fix user variables in template acl strings
-    ACLStringIterator.next = monkey_patch(ACLStringIterator.next,
-                                          acl_user_expand)
 
     _hooks_installed = True
